+++ common/src/main/java/net/more_rpg_classes/effect/FatalPoisonEffect.java
@@ -5,7 +5,6 @@ import net.minecraft.entity.EntityType;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.effect.StatusEffect;
 import net.minecraft.entity.effect.StatusEffectCategory;
-import net.more_rpg_classes.client.particle.MoreParticles;
 import net.more_rpg_classes.damage.PoisonDamageSource;
 import net.more_rpg_classes.util.tags.MRPGCEntityTags;
 
@@ -18,7 +17,7 @@ public class FatalPoisonEffect extends StatusEffect {
     public  boolean applyUpdateEffect(LivingEntity entity, int amplifier) {
         EntityType<?> type = ((Entity) entity).getType();
         if(!type.isIn(MRPGCEntityTags.POISON_IMMUNE)){
-            entity.damage(new PoisonDamageSource(entity.getDamageSources().magic().getTypeRegistryEntry()), 1F);
+            entity.damage(new PoisonDamageSource(entity.getDamageSources().magic().getTypeRegistryEntry()), 1F + amplifier);
         }
         return true;
     }
@@ -26,12 +25,7 @@ public class FatalPoisonEffect extends StatusEffect {
 
     @Override
     public boolean canApplyUpdateEffect(int duration, int amplifier) {
-        int i = 25 >> amplifier;
-        if (i > 0) {
-            return duration % i == 0;
-        } else {
-            return true;
-        }
+        return duration % 25 == 0;
     }
 
 }